requirements.txt
python-dotenv
langgraph
langchain
langchain_community
langchain_ollama
langchain_openai



# langchain
# langchain_core
# langchain_community
# langchain_huggingface
# langchain_chroma
# langchain_experimental
# langchain_ollama
# beautifulsoup4
# ollama